[network-data] add check for prefix length

+++ src/core/common/encoding.hpp
@@ -76,7 +76,7 @@ inline uint32_t Reverse32(uint32_t v)
     return v;
 }
 
-#define BitVectorBytes(x) (((x) + (CHAR_BIT - 1)) / CHAR_BIT)
+#define BitVectorBytes(x) static_cast<uint8_t>(((x) + (CHAR_BIT - 1)) / CHAR_BIT)
 
 namespace BigEndian {
